Kia: update required packages (#25824)
* Kia: LKAS is standard on Ceed 2019

https://www.downeys.co.uk/newmodels/Ceed_V3_24_09_2019.pdf

* Kia: delete Forte 2018

It doesn't appear that this vehicle has Adaptive/Smart Cruise Control in
any form, which was later added in 2019.

https://cdn.dealereprocess.org/cdn/brochures/kia/2018-forte.pdf
https://cdn.dealereprocess.org/cdn/brochures/kia/2019-forte.pdf

* Kia: LKAS is standard on Forte 2019+

https://cdn.dealereprocess.org/cdn/brochures/kia/2019-forte.pdf
https://cdn.dealereprocess.org/cdn/brochures/kia/2020-forte.pdf

* Kia: rename Niro Electric to EV

https://www.kia.com/us/en/niro

Co-authored-by: Shane Smiskol <shane@smiskol.com>

* Kia: LKAS is standard on Niro PHEV 2018+

https://cdn.dealereprocess.org/cdn/brochures/kia/2018-niro.pdf
https://cdn.dealereprocess.org/cdn/brochures/kia/2019-niro.pdf

* Kia: update required package on Optima 2017

The ACC package on the Optima 2017 is named "Advanced Smart Cruise
Control". It also doesn't have an LKAS package, only LDWS is available.

https://cdn.dealereprocess.org/cdn/brochures/kia/2017-optima.pdf

Co-authored-by: Shane Smiskol <shane@smiskol.com>

* Kia: LKAS is standard on Optima 2019

https://cdn.dealereprocess.org/cdn/brochures/kia/2019-optima.pdf

* Kia: revert package change to Seltos 2021

LKAS is NOT a standard package on the Seltos 2021

https://cdn.dealereprocess.org/cdn/brochures/kia/2021-seltos.pdf

* Kia: update required package on Sorento 2018

Similar to the Optima 2017, the ACC package on the Sorento 2018 is named
"Advanced Smart Cruise Control". It also doesn't have an LKAS package,
only LDWS.

SCC and LKAS were introduced in MY2019.

https://cdn.dealereprocess.org/cdn/brochures/kia/2018-sorento.pdf

Co-authored-by: Shane Smiskol <shane@smiskol.com>

* Kia: SCC is standard on Kia Niro PHEV 2018-19

https://cdn.dealereprocess.org/cdn/brochures/kia/2018-niro.pdf

Co-authored-by: Shane Smiskol <shane@smiskol.com>

* Kia: update required package on Optima Hybrid 2017

Similar to the Optima 2017, the ACC package on he Optima Hybrid 2017 is
named "Advanced Smart Cruise Control". It also doesn't have an LKAS
pacakge, only LDWS.

https://cdn.dealereprocess.org/cdn/brochures/kia/2017-optimahybrid.pdf

* update docs

Co-authored-by: Shane Smiskol <shane@smiskol.com>
old-commit-hash: 06fb52c146aa9e00126127ac28fbbd22e9914e6f